Add competition to find ratings for all bots
Main.py is now intended for single match verbose debugging. If you want
to test a bot against the entire field, use competition.py. Ratings
require the trueskill module which can be installed with "easy_install
trueskill" from a command prompt if python is in your path, or with the
same command from the relevant python directory if it's not.